fix(autobahn): widen inboundFullnodeCap to int64 to satisfy gosec
The 10000 bound makes the int→int32 cast safe by construction,
but gosec's G115 check is flow-insensitive and can't see across
buildDataState's validation. Switching the counter (atomic.Int64)
and cap (int64) eliminates the cast entirely.

+++ sei-tendermint/internal/p2p/giga_router_fullnode.go
@@ -32,7 +32,7 @@ func NewGigaFullnodeRouter(cfg *GigaRouterCommonConfig, key NodeSecretKey) (*gig
 			poolIn:             giga.NewPool[NodePublicKey, rpc.Server[giga.API]](),
 			poolOut:            giga.NewPool[NodePublicKey, rpc.Client[giga.API]](),
 			app:                cfg.App,
-			inboundFullnodeCap: int32(cfg.MaxInboundFullnodePeers),
+			inboundFullnodeCap: int64(cfg.MaxInboundFullnodePeers),
 		},
 	}, nil
 }
